Fix scalar methods to treat NAN correctlt
