+++ lib/x86simdsort-scalar.h
@@ -1,36 +1,40 @@
 #include <algorithm>
 #include <numeric>
-#define UNUSED(x) (void)(x)
+#include "custom-compare.h"
+
 namespace xss {
 namespace scalar {
-    /* TODO: handle NAN */
     template <typename T>
     void qsort(T *arr, int64_t arrsize)
     {
-        std::sort(arr, arr + arrsize);
+        std::sort(arr, arr + arrsize, compare<T, std::less<T>>());
     }
     template <typename T>
     void qselect(T *arr, int64_t k, int64_t arrsize, bool hasnan)
     {
-        UNUSED(hasnan);
-        std::nth_element(arr, arr + k, arr + arrsize);
+        if (hasnan) {
+            std::nth_element(arr, arr + k, arr + arrsize, compare<T, std::less<T>>());
+        }
+        else {
+            std::nth_element(arr, arr + k, arr + arrsize);
+        }
     }
     template <typename T>
     void partial_qsort(T *arr, int64_t k, int64_t arrsize, bool hasnan)
     {
-        UNUSED(hasnan);
-        std::partial_sort(arr, arr + k, arr + arrsize);
+        if (hasnan) {
+            std::partial_sort(arr, arr + k, arr + arrsize, compare<T, std::less<T>>());
+        }
+        else {
+            std::partial_sort(arr, arr + k, arr + arrsize);
+        }
     }
     template <typename T>
     std::vector<int64_t> argsort(T *arr, int64_t arrsize)
     {
         std::vector<int64_t> arg(arrsize);
         std::iota(arg.begin(), arg.end(), 0);
-        std::sort(arg.begin(),
-                  arg.end(),
-                  [arr](int64_t left, int64_t right) -> bool {
-                      return arr[left] < arr[right];
-                  });
+        std::sort(arg.begin(), arg.end(), compare_arg<T, std::less<T>>(arr));
         return arg;
     }
     template <typename T>
@@ -41,9 +45,7 @@ namespace scalar {
         std::nth_element(arg.begin(),
                          arg.begin() + k,
                          arg.end(),
-                         [arr](int64_t left, int64_t right) -> bool {
-                             return arr[left] < arr[right];
-                         });
+                         compare_arg<T, std::less<T>>(arr));
         return arg;
     }
